Split Kokkos_CopyViews.hpp
