+++ core/src/Kokkos_Core.hpp
@@ -75,7 +75,10 @@
 #include <Kokkos_Tuners.hpp>
 #include <Kokkos_TaskScheduler.hpp>
 #include <Kokkos_Complex.hpp>
-#include <Kokkos_CopyViews.hpp>
+#include <View/Kokkos_DeepCopy.hpp>
+#include <View/Kokkos_CreateMirror.hpp>
+#include <View/Kokkos_ResizeRealloc.hpp>
+#include <View/Kokkos_LocalDeepCopy.hpp>
 #include <impl/Kokkos_InitializationSettings.hpp>
 #include <functional>
 #include <iosfwd>
